extract inline tests second pass
